Fixed AES secret key setter from Config class
Transaction amount is now a double in the database
Added getFromAbbreviation to TransactionType and Currency
Moved properties from Core Config to API Config
Location returns campus coordinates when zero

+++ src/main/resources/db/migration/V1__tables.sql
@@ -36,7 +36,7 @@ CREATE TABLE IF NOT EXISTS `transaction` (
   `id` INT NOT NULL,
   `title` VARCHAR(45) NOT NULL,
   `description` VARCHAR(255) NULL,
-  `amount` INT NOT NULL,
+  `amount` DOUBLE NOT NULL,
   `currency` VARCHAR(3) NOT NULL,
   `type` VARCHAR(7) NOT NULL,
   `timestamp` DATETIME NOT NULL,
